Return block size log2 as uint8_t

+++ src/file.cpp
@@ -40,8 +40,7 @@ class File::DataCategoryReader {
   virtual size_t GetMetadataSize() const = 0;
   virtual size_t GetMetadataItemsCount() const {
     return FileLayout::MetadataItemsCount(FileLayout::CategoryFromValue(file_->metadata()->size_category.value()),
-                                          file_->metadata()->size_on_disk.value(),
-                                          static_cast<uint8_t>(file_->quota()->block_size_log2()));
+                                          file_->metadata()->size_on_disk.value(), file_->quota()->block_size_log2());
   }
 
   virtual std::span<const std::byte> GetData(size_t offset, size_t size) = 0;
@@ -290,7 +289,7 @@ class File::DataCategory4Reader : public File::DataCategory3Reader {
   }
 
   size_t ClustersInBlock() const {
-    return FileLayout::ClustersPerClusterMetadataBlock(static_cast<uint8_t>(file_->quota()->block_size_log2()));
+    return FileLayout::ClustersPerClusterMetadataBlock(file_->quota()->block_size_log2());
   }
 };
